+++ src/lib/components/NavBar.svelte
@@ -104,13 +104,16 @@
             <div
               class="z-10 overflow-hidden rounded-md bg-zinc-900/80 px-1 text-xs text-nowrap text-ellipsis backdrop-blur-xs"
             >
-              {#each audioPlayer.currentTrack.artists.sort( (a, _) => (a.name !== audioPlayer.currentTrack?.album.albumArtist.name ? 1 : -1) ) as artist, index (artist.id)}
+              {#each audioPlayer.currentTrack.artists.toSorted( (a, _) => (a.name !== audioPlayer.currentTrack?.album.albumArtist.name ? 1 : -1) ) as artist, index (artist.id)}
+                {@const shouldHaveComma =
+                  audioPlayer.currentTrack.artists.length > 1 &&
+                  index != audioPlayer.currentTrack.artists.length - 1}
                 <button
                   title={artist.name}
-                  class="hover:underline"
+                  class={['hover:underline', shouldHaveComma && 'mr-1']}
                   onclick={() => goto(`/artist/${artist.id}`)}
                 >
-                  {artist.name}{#if audioPlayer.currentTrack.artists.length > 1 && index != audioPlayer.currentTrack.artists.length - 1},{/if}
+                  {artist.name}{#if shouldHaveComma},{/if}
                 </button>
               {/each}
             </div>
@@ -138,7 +141,7 @@
 </div>
 
 <div
-  class="absolute top-0 z-40 flex h-[calc(100%-11.25rem)] justify-center overflow-x-clip overflow-y-auto rounded-md bg-zinc-900/80 backdrop-blur-md transition-all duration-300 sm:hidden"
+  class="absolute top-0 z-40 flex h-[calc(100%-11.25rem)] justify-center overflow-x-clip overflow-y-auto rounded-xl bg-zinc-900/80 backdrop-blur-md transition-all duration-300 sm:hidden"
   class:w-full={open}
   class:left-0={open}
   class:w-0={!open}