+++ src/service-worker.ts
@@ -1,5 +1,7 @@
-/// <reference types="@sveltejs/kit" />
+/// <reference no-default-lib="true"/>
+/// <reference lib="esnext" />
 /// <reference lib="webworker" />
+/// <reference types="@sveltejs/kit" />
 
 declare let self: ServiceWorkerGlobalScope;